Reintroduced the "Hotspot Feature" for custom apps.
