+++ app/src/main/java/org/kiwix/kiwixmobile/main/KiwixMainActivity.kt
@@ -23,7 +23,6 @@ import android.content.res.Configuration
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
-import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.view.ActionMode
 import androidx.core.content.pm.ShortcutInfoCompat
@@ -46,7 +45,6 @@ import kotlinx.coroutines.launch
 import org.kiwix.kiwixmobile.BuildConfig
 import org.kiwix.kiwixmobile.R
 import org.kiwix.kiwixmobile.core.R.drawable
-import org.kiwix.kiwixmobile.core.R.id
 import org.kiwix.kiwixmobile.core.R.mipmap
 import org.kiwix.kiwixmobile.core.R.string
 import org.kiwix.kiwixmobile.core.base.FragmentActivityExtensions
@@ -107,6 +105,8 @@ class KiwixMainActivity : CoreMainActivity() {
   override val notesFragmentResId: Int = R.id.notesFragment
   override val readerFragmentResId: Int = R.id.readerFragment
   override val helpFragmentResId: Int = R.id.helpFragment
+  override val zimHostFragmentResId: Int = R.id.zimHostFragment
+  override val navGraphId: Int = R.navigation.kiwix_nav_graph
   override val topLevelDestinations =
     setOf(R.id.downloadsFragment, R.id.libraryFragment, R.id.readerFragment)
 
@@ -308,19 +308,6 @@ class KiwixMainActivity : CoreMainActivity() {
         OPENING_ZIM_FILE_DELAY
       )
     }
-  }
-
-  override fun onNavigationItemSelected(item: MenuItem): Boolean {
-    when (item.itemId) {
-      id.menu_host_books -> openZimHostFragment()
-      else -> return super.onNavigationItemSelected(item)
-    }
-    return true
-  }
-
-  private fun openZimHostFragment() {
-    disableDrawer()
-    navigate(R.id.zimHostFragment)
   }
 
   override fun getIconResId() = mipmap.ic_launcher
