+++ app/src/main/res/navigation/kiwix_nav_graph.xml
@@ -125,7 +125,7 @@
     tools:layout="@layout/activity_language" />
   <fragment
     android:id="@+id/zimHostFragment"
-    android:name="org.kiwix.kiwixmobile.webserver.ZimHostFragment"
+    android:name="org.kiwix.kiwixmobile.core.webserver.ZimHostFragment"
     android:label="ZimHostFragment"
     tools:layout="@layout/activity_zim_host" />
   <fragment
