+++ core/src/main/java/org/kiwix/kiwixmobile/core/webserver/ZimHostCallbacks.kt
@@ -1,6 +1,6 @@
 /*
  * Kiwix Android
- * Copyright (c) 2023 Kiwix <android.kiwix.org>
+ * Copyright (c) 2019 Kiwix <android.kiwix.org>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
@@ -15,10 +15,12 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
+package org.kiwix.kiwixmobile.core.webserver
 
-package org.kiwix.kiwixmobile.webserver.wifi_hotspot
-
-data class ServerStatus(
-  val isServerStarted: Boolean,
-  val errorMessage: Int? = null
-)
+interface ZimHostCallbacks {
+  fun onServerStarted(ip: String)
+  fun onServerStopped()
+  fun onServerFailedToStart(errorMessage: Int?)
+  fun onIpAddressValid()
+  fun onIpAddressInvalid()
+}
