+++ core/src/main/java/org/kiwix/kiwixmobile/core/webserver/wifi_hotspot/HotspotStateReceiver.kt
@@ -15,12 +15,12 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package org.kiwix.kiwixmobile.webserver.wifi_hotspot
+package org.kiwix.kiwixmobile.core.webserver.wifi_hotspot
 
 import android.content.Context
 import android.content.Intent
 import org.kiwix.kiwixmobile.core.base.BaseBroadcastReceiver
-import org.kiwix.kiwixmobile.webserver.wifi_hotspot.HotspotStateReceiver.HotspotState.DISABLED
+import org.kiwix.kiwixmobile.core.webserver.wifi_hotspot.HotspotStateReceiver.HotspotState.DISABLED
 import javax.inject.Inject
 
 const val EXTRA_WIFI_AP_STATE = "wifi_state"
