+++ core/src/main/res/layout/activity_zim_host.xml
